requirements.txt
fastapi
celery>=5.3.0
pydantic[email]
uvicorn
pydantic-settings
python-multipart
sqlalchemy>=2.0.23
PyYAML
python-dotenv
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.12.1
numpy
httpx>=0.24.0
aiofiles
redis==5.2.1